fix: Fixed Header and backend, removed unnecessary route

+++ back-end/src/database/migrations/20221129183931-products.js
@@ -13,7 +13,7 @@ module.exports = {
           allowNull: false,
         },
         price: {
-          type: Sequelize.INTEGER,
+          type: Sequelize.DECIMAL(4,2),
           allowNull: false,
         },
         url_image: {
